further refactors of server-side example
