+++ examples/zkapps/interacting-with-zkApps-server-side/.eslintrc.cjs
@@ -5,11 +5,19 @@ module.exports = {
     node: true,
     jest: true,
   },
-  extends: ['eslint:recommended', 'plugin:snarkyjs/recommended'],
+  extends: [
+    'eslint:recommended',
+    'plugin:@typescript-eslint/eslint-recommended',
+    'plugin:@typescript-eslint/recommended',
+    'plugin:snarkyjs/recommended',
+  ],
   parser: '@typescript-eslint/parser',
   parserOptions: {
     ecmaVersion: 'latest',
   },
   plugins: ['@typescript-eslint', 'snarkyjs'],
-  rules: {},
+  rules: {
+    'no-constant-condition': 'off',
+    'prefer-const': 'off',
+  },
 };
