--- examples/zkapps/interacting-with-zkApps-server-side/src/deploy.ts
@@ -1,63 +1,0 @@
-import { Square } from './Square.js';
-import { Field, Mina, PrivateKey, AccountUpdate, fetchAccount } from 'snarkyjs';
-
-export { deploy };
-
-const transactionFee = 100_000_000;
-
-async function deploy(
-  deployerPrivateKey: PrivateKey,
-  zkAppPrivateKey: PrivateKey,
-  zkapp: Square,
-  verificationKey: { data: string; hash: string | Field }
-) {
-  let sender = deployerPrivateKey.toPublicKey();
-  let zkAppPublicKey = zkAppPrivateKey.toPublicKey();
-  console.log('using deployer private key with public key', sender.toBase58());
-  console.log(
-    'using zkApp private key with public key',
-    zkAppPublicKey.toBase58()
-  );
-
-  // ----------------------------------------------------
-
-  let { account } = await fetchAccount({ publicKey: zkAppPublicKey });
-  let isDeployed = account?.verificationKey !== undefined;
-
-  // ----------------------------------------------------
-
-  if (isDeployed) {
-    console.log(
-      'zkApp for public key',
-      zkAppPublicKey.toBase58(),
-      'found deployed'
-    );
-  } else {
-    console.log('Deploying zkapp for public key', zkAppPublicKey.toBase58());
-    let transaction = await Mina.transaction(
-      { sender, fee: transactionFee },
-      () => {
-        AccountUpdate.fundNewAccount(sender);
-        zkapp.deploy({ verificationKey });
-      }
-    );
-    await transaction.prove();
-    transaction.sign([deployerPrivateKey, zkAppPrivateKey]);
-
-    console.log('Sending the deploy transaction...');
-    const res = await transaction.send();
-    const hash = res.hash();
-    if (hash === undefined) {
-      console.log('error sending transaction (see above)');
-    } else {
-      console.log(
-        'See deploy transaction at',
-        'https://berkeley.minaexplorer.com/transaction/' + hash
-      );
-      console.log('waiting for zkApp account to be deployed...');
-      await res.wait();
-      isDeployed = true;
-    }
-  }
-  return isDeployed;
-}
