+++ examples/zkapps/interacting-with-zkApps-server-side/src/main.ts
@@ -1,13 +1,8 @@
 import { Square } from './Square.js';
 import { isReady, shutdown, Mina, PrivateKey } from 'snarkyjs';
 
-import { deploy } from './deploy.js';
 import fs from 'fs';
-import {
-  loopUntilAccountExists,
-  makeAndSendTransaction,
-  accountExists,
-} from './utils.js';
+import { loopUntilAccountExists, deploy } from './utils.js';
 
 await isReady;
 
@@ -64,26 +59,57 @@ let { verificationKey } = await Square.compile();
 const zkAppPublicKey = zkAppPrivateKey.toPublicKey();
 let zkapp = new Square(zkAppPublicKey);
 
-let accountExistsAlready = await accountExists(zkAppPublicKey);
-if (!accountExistsAlready) {
-  // Programmatic deploy:
-  //   Besides the CLI, you can also create accounts programmatically. This is useful if you need
-  //   more custom account creation - say deploying a zkApp to a different key than the fee payer
-  //   key, programmatically parameterizing a zkApp before initializing it, or creating Smart
-  //   Contracts programmatically for users as part of an application.
-  await deploy(deployerPrivateKey, zkAppPrivateKey, zkapp, verificationKey);
-}
+// Programmatic deploy:
+//   Besides the CLI, you can also create accounts programmatically. This is useful if you need
+//   more custom account creation - say deploying a zkApp to a different key than the fee payer
+//   key, programmatically parameterizing a zkApp before initializing it, or creating Smart
+//   Contracts programmatically for users as part of an application.
+await deploy(deployerPrivateKey, zkAppPrivateKey, zkapp, verificationKey);
+
+await loopUntilAccountExists({
+  account: zkAppPublicKey,
+  eachTimeNotExist: () =>
+    console.log('waiting for zkApp account to be deployed...'),
+  isZkAppAccount: true,
+});
 
 let num = (await zkapp.num.fetch())!;
 console.log(`current value of num is ${num}`);
 
 // ----------------------------------------------------
 
-await makeAndSendTransaction({
-  feePayerPrivateKey: deployerPrivateKey,
-  mutateZkApp: () => zkapp.update(num.mul(num)),
-  transactionFee,
-});
+let transaction = await Mina.transaction(
+  { sender: deployerPublicKey, fee: transactionFee },
+  () => {
+    zkapp.update(num.mul(num));
+  }
+);
+
+// fill in the proof - this can take a while...
+console.log('Creating an execution proof...');
+let time0 = performance.now();
+await transaction.prove();
+let time1 = performance.now();
+console.log(`creating proof took ${(time1 - time0) / 1e3} seconds`);
+
+// sign transaction with the deployer account
+transaction.sign([deployerPrivateKey]);
+
+console.log('Sending the transaction...');
+let pendingTransaction = await transaction.send();
+
+// ----------------------------------------------------
+
+if (!pendingTransaction.isSuccess) {
+  console.log('error sending transaction (see above)');
+  process.exit(0);
+}
+
+console.log(
+  `See transaction at https://berkeley.minaexplorer.com/transaction/${pendingTransaction.hash()}
+Waiting for transaction to be included...`
+);
+await pendingTransaction.wait();
 
 console.log(`updated state! ${await zkapp.num.fetch()}`);
