+++ examples/zkapps/interacting-with-zkApps-server-side/src/utils.ts
@@ -1,16 +1,14 @@
-import { Mina, PrivateKey, PublicKey, fetchAccount } from 'snarkyjs';
+import {
+  PublicKey,
+  fetchAccount,
+  PrivateKey,
+  Field,
+  Mina,
+  AccountUpdate,
+} from 'snarkyjs';
+import { Square } from './Square';
 
-export { accountExists, loopUntilAccountExists, makeAndSendTransaction };
-
-// ========================================================
-
-async function accountExists(account: PublicKey) {
-  let response = await fetchAccount({ publicKey: account });
-  let accountExists = response.account !== undefined;
-  return accountExists;
-}
-
-// ========================================================
+export { loopUntilAccountExists, deploy };
 
 async function loopUntilAccountExists({
   account,
@@ -37,44 +35,58 @@ async function loopUntilAccountExists({
   }
 }
 
-// ========================================================
+const deployTransactionFee = 100_000_000;
 
-async function makeAndSendTransaction({
-  feePayerPrivateKey,
-  mutateZkApp,
-  transactionFee,
-}: {
-  feePayerPrivateKey: PrivateKey;
-  mutateZkApp: () => void;
-  transactionFee: number;
-}) {
-  const sender = feePayerPrivateKey.toPublicKey();
-
-  let transaction = await Mina.transaction(
-    { sender, fee: transactionFee },
-    () => {
-      mutateZkApp();
-    }
+async function deploy(
+  deployerPrivateKey: PrivateKey,
+  zkAppPrivateKey: PrivateKey,
+  zkapp: Square,
+  verificationKey: { data: string; hash: string | Field }
+) {
+  let sender = deployerPrivateKey.toPublicKey();
+  let zkAppPublicKey = zkAppPrivateKey.toPublicKey();
+  console.log('using deployer private key with public key', sender.toBase58());
+  console.log(
+    'using zkApp private key with public key',
+    zkAppPublicKey.toBase58()
   );
 
-  // fill in the proof - this can take a while...
-  console.log('Creating an execution proof...');
-  const time0 = Date.now();
-  await transaction.prove();
-  const time1 = Date.now();
-  console.log('creating proof took', (time1 - time0) / 1e3, 'seconds');
+  let { account } = await fetchAccount({ publicKey: zkAppPublicKey });
+  let isDeployed = account?.verificationKey !== undefined;
 
-  console.log('Sending the transaction...');
-  const res = await transaction.sign([feePayerPrivateKey]).send();
-  const hash = res.hash();
-  if (hash === undefined) {
-    console.log('error sending transaction (see above)');
-  } else {
+  if (isDeployed) {
     console.log(
-      'See transaction at',
-      'https://berkeley.minaexplorer.com/transaction/' + hash
+      'zkApp for public key',
+      zkAppPublicKey.toBase58(),
+      'found deployed'
     );
-    console.log('Waiting for transaction to be included...');
-    await res.wait();
+  } else {
+    console.log('Deploying zkapp for public key', zkAppPublicKey.toBase58());
+    let transaction = await Mina.transaction(
+      { sender, fee: deployTransactionFee },
+      () => {
+        AccountUpdate.fundNewAccount(sender);
+        // NOTE: this calls `init()` if this is the first deploy
+        zkapp.deploy({ verificationKey });
+      }
+    );
+    await transaction.prove();
+    transaction.sign([deployerPrivateKey, zkAppPrivateKey]);
+
+    console.log('Sending the deploy transaction...');
+    const res = await transaction.send();
+    const hash = res.hash();
+    if (hash === undefined) {
+      console.log('error sending transaction (see above)');
+    } else {
+      console.log(
+        'See deploy transaction at',
+        'https://berkeley.minaexplorer.com/transaction/' + hash
+      );
+      console.log('waiting for zkApp account to be deployed...');
+      await res.wait();
+      isDeployed = true;
+    }
   }
+  return isDeployed;
 }
